requirements.txt
# Core dependencies
langchain
langchain-openai
langchain-community
langchain-core
langgraph

# Database
psycopg2-binary
sqlalchemy==2.0.25

# Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Data processing
pandas
numpy
openpyxl==3.1.2
xlrd==2.0.1

# API and web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Task Queue & Message Broker
celery[redis]==5.3.4
redis>=4.5.2,<5.0.0
flower==2.0.1

# Authentication
google-auth
google-auth-oauthlib
google-auth-httplib2
python-jose[cryptography]
PyJWT
passlib[bcrypt]
authlib
httpx
